refactor: 인스턴스 컴포넌트 구조 개선 및 기능별 분리
