+++ src/components/instance/create/FlavorSelect.vue
